requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
redis>=5.0.7
qdrant-client>=1.9.1
langchain>=0.2.10
langchain-core>=0.2.26
langchain-community>=0.2.9
langchain-openai>=0.1.20
openai>=1.37.0
pydantic>=2.8.2
